三角形周长和面积（整数与双精度）

+++ targ_CS_and_SS_double_ver.c
@@ -2,13 +2,13 @@
 #include<math.h>
 
 int main(){
-	int a,b,c;
-	int CS,SS;
+	double a,b,c;
+	double CS,SS;
 	printf("请输入三角形三边长(整数)，用空格隔开：");
-	scanf("%d %d %d",&a,&b,&c);
+	scanf("%lf %lf %lf",&a,&b,&c);
 	CS = a + b + c;
-	printf("三角形周长：%d\n",CS);
+	printf("三角形周长：%lf\n",CS);
 	SS = sqrt(CS/2 * (CS/2 - a) * (CS/2 - b) * (CS/2 - c));
-	printf("三角形面积：%d\n",SS);
+	printf("三角形面积：%lf\n",SS);
 	return 0;
 }
